+++ src/components/base/BasePage.js
@@ -1,0 +1,8 @@
+import React, { Component } from 'react';
+
+export default class BasePage extends Component{
+
+    showMessageError(message){
+        this.setState({openMessage: true,errorMessage:message});
+    }
+}
