+++ src/components/pages/NotFoundPage.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
-import Header from './Header';
-import '../resources/styles/index.css';
+import Header from '../base/Header';
+import '../../resources/styles/index.css';
 
 class NotFound extends Component{
     render(){
         return (
             <div>
                 <Header />
-                <center><img alt="" src={require("../resources/images/not_found.png")} /></center>
+                <center><img alt="" src={require("../../resources/images/not_found.png")} /></center>
             </div>
         );
     }
